# Add "project" mode for faster code coverage and allocation tracking