--- a/doc/man/julia.1
+++ b/doc/man/julia.1
@@ -195,11 +195,11 @@ Generate LLVM bitcode (.bc)
 Generate an incremental output file (rather than complete)
 
 .TP
---code-coverage={none|user|all}, --code-coverage
+--code-coverage={none|project|user|all}, --code-coverage
 Count executions of source lines (omitting setting is equivalent to 'user')
 
 .TP
---track-allocation={none|user|all}, --track-allocation
+--track-allocation={none|project|user|all}, --track-allocation
 Count bytes allocated by each source line
 
 .SH FILES