@DilumAluthge 2022-01-08
Presumably, a more general solution would be to allow e.g. `--code-coverage=path` and `--code-coverage-path=/foo/bar/baz`? As you originally suggested in #41626?