@IanButterworth 2022-01-08
That's already reserved for specifying a file to save the coverage report to, unfortunately