@DilumAluthge 2022-01-08
Maybe something like `--code-coverage=trackpaths` and `--code-coverage-trackpaths=/foo/bar/baz`?