@IanButterworth 2022-01-08
Are you suggesting these because you think it's better than what's here?

Note that I just pushed automatically setting the "project module" and updated the description up top