@DilumAluthge 2022-01-08
I don't feel strongly either way. It's just that I liked your original suggestion in https://github.com/JuliaLang/julia/issues/41626#issue-946896384, to be able to set a path (`--code-coverage=/path/to/repo`).

And I was wondering if there would be a use case where people would have a couple of different packages that they want to get code coverage for at the same time.

So I was thinking that your original suggestion of `--code-coverage=/path/to/repo` could be used if e.g. people `dev` all of those packages into a single parent directory, and then enable code coverage on all modules in that parent directory.