@DilumAluthge 2022-01-08
The counter-point, of course, is that we only need to worry about implementing the use case that 95% of people will have, which I think is covered by this PR. And then people that have more complex needs will just have to use `--code-coverage=user` and take the additional slowdown, which is probably acceptable.