@IanButterworth 2022-01-08
> a couple of different packages that they want to get code coverage for at the same time.

That's possible here via
```
push!(Base.explicit_project_dirs, dir1, dir2)
```