@DilumAluthge 2022-01-08
I was hoping to have that exposed on the command-line, because then we could consider having an `input` to the `julia-runtest` GitHub Action.

But again, it's probably fine to tell users that need that functionality to just write out their test commands explicitly.