@IanButterworth 2022-01-08
I think as long as people can directly set `--code-coverage=all` in `julia-runtest` to override the Pkg setting, this will work in 99% of cases out of the box, and also work in those others.. I think.

I just added a benchmark example up top. It's a little hard to find a package with a big stack to test because there are some breakages in master currently