@IanButterworth 2022-01-25
@JeffBezanson do you think this is a good idea? If so, I'd be happy to fix up tests etc.