@IanButterworth 2022-02-03
Triage suggested simplifying this instead with a `--code-coverage=@ModuleName` api and checking `Base.moduleroot(mod)` against that